Add set_now() to update local cache immediatly when remote storage is slow
set() never updates the local config cache. It sends the update to the
storage engine, and it is the storage engine who eventually updates
the local config cache.

This makes sure all local caches are not updated until the remote config
service is notified and responds. On the other hand, a set(k => v)
followed by get(k) might not return the v.

On some situations you do want to update the local cache immediatly. That's
what set_now() does. It updates the local cache and sends the update to the
storage engine.

Notifications are only sent when the remote engine is updated and acks the
new config, even when set_now() is used.

Signed-off-by: Pedro Melo <melo@simplicidade.org>