Redo previous fix to make sure Storage init() is called:
The Singleton role provides a init() hook point that other roles can use
to perform initialization.

The Storage roles does that, making sure storage init() method is called.

This fix is better because there was no garantee that storage init() was
called before any cache accesses on the previous version.

Signed-off-by: Pedro Melo <melo@simplicidade.org>